# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DaliJob API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/profile/resume-imports": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Import Resume Pdf",
        "operationId": "import_resume_pdf_api_v1_profile_resume_imports_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_resume_pdf_api_v1_profile_resume_imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/resume-imports/{document_id}/retry": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Retry Resume Import",
        "operationId": "retry_resume_import_api_v1_profile_resume_imports__document_id__retry_post",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/resume-imports/apply": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Apply Resume Import",
        "operationId": "apply_resume_import_api_v1_profile_resume_imports_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeImportApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resume-profiles": {
      "get": {
        "tags": [
          "resume-profiles"
        ],
        "summary": "List Resume Profiles",
        "operationId": "list_resume_profiles_api_v1_resume_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeProfileListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resume-profiles"
        ],
        "summary": "Create Resume Profile",
        "operationId": "create_resume_profile_api_v1_resume_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeProfileCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resume-profiles/{resume_profile_id}": {
      "get": {
        "tags": [
          "resume-profiles"
        ],
        "summary": "Get Resume Profile",
        "operationId": "get_resume_profile_api_v1_resume_profiles__resume_profile_id__get",
        "parameters": [
          {
            "name": "resume_profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resume Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "resume-profiles"
        ],
        "summary": "Update Resume Profile",
        "operationId": "update_resume_profile_api_v1_resume_profiles__resume_profile_id__patch",
        "parameters": [
          {
            "name": "resume_profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resume Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "resume-profiles"
        ],
        "summary": "Delete Resume Profile",
        "operationId": "delete_resume_profile_api_v1_resume_profiles__resume_profile_id__delete",
        "parameters": [
          {
            "name": "resume_profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resume Profile Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resume-profiles/{resume_profile_id}/dependencies": {
      "get": {
        "tags": [
          "resume-profiles"
        ],
        "summary": "Get Resume Profile Dependencies",
        "operationId": "get_resume_profile_dependencies_api_v1_resume_profiles__resume_profile_id__dependencies_get",
        "parameters": [
          {
            "name": "resume_profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resume Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeProfileDependencyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_resume_pdf_api_v1_profile_resume_imports_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_resume_pdf_api_v1_profile_resume_imports_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumeData": {
        "properties": {
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "experience": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Experience"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "education": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Education"
          },
          "certifications": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Certifications"
          },
          "projects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Projects"
          },
          "awards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Awards"
          },
          "publications": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Publications"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "volunteer": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Volunteer"
          },
          "target_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Roles"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ResumeData"
      },
      "ResumeImportApplyRequest": {
        "properties": {
          "resume_data": {
            "$ref": "#/components/schemas/ResumeData"
          },
          "source_document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Id"
          },
          "source_document_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Version Id"
          }
        },
        "type": "object",
        "required": [
          "resume_data"
        ],
        "title": "ResumeImportApplyRequest"
      },
      "ResumeImportResponse": {
        "properties": {
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "document_version_id": {
            "type": "integer",
            "title": "Document Version Id"
          },
          "extracted_text_preview": {
            "type": "string",
            "title": "Extracted Text Preview"
          },
          "suggestions": {
            "$ref": "#/components/schemas/ResumeData"
          },
          "parse_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Warning"
          }
        },
        "type": "object",
        "required": [
          "file_name",
          "document_id",
          "document_version_id",
          "extracted_text_preview",
          "suggestions"
        ],
        "title": "ResumeImportResponse"
      },
...
      "ResumeProfileCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "default": "Master Resume"
          },
          "resume_data": {
            "$ref": "#/components/schemas/ResumeData"
          },
          "source_document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Id"
          },
          "source_document_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Version Id"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          }
        },
        "type": "object",
        "title": "ResumeProfileCreateRequest"
      },
      "ResumeProfileDependency": {
        "properties": {
          "dependency_type": {
            "type": "string",
            "title": "Dependency Type"
          },
          "dependency_count": {
            "type": "integer",
            "title": "Dependency Count"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "dependency_type",
          "dependency_count",
          "message"
        ],
        "title": "ResumeProfileDependency"
      },
      "ResumeProfileDependencyResponse": {
        "properties": {
          "can_delete_without_warning": {
            "type": "boolean",
            "title": "Can Delete Without Warning"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/ResumeProfileDependency"
            },
            "type": "array",
            "title": "Dependencies"
          }
        },
        "type": "object",
        "required": [
          "can_delete_without_warning"
        ],
        "title": "ResumeProfileDependencyResponse"
      },
      "ResumeProfileListResponse": {
        "properties": {
          "resume_profiles": {
            "items": {
              "$ref": "#/components/schemas/ResumeProfileResponse"
            },
            "type": "array",
            "title": "Resume Profiles"
          }
        },
        "type": "object",
        "required": [
          "resume_profiles"
        ],
        "title": "ResumeProfileListResponse"
      },
      "ResumeProfileResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "workspace_id": {
            "type": "integer",
            "title": "Workspace Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "resume_data": {
            "$ref": "#/components/schemas/ResumeData"
          },
          "source_document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Id"
          },
          "source_document_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Version Id"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "user_id",
          "title",
          "resume_data",
          "is_default",
          "created_at",
          "updated_at"
        ],
        "title": "ResumeProfileResponse"
      },
      "ResumeProfileUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "resume_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResumeData"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          }
        },
        "type": "object",
        "title": "ResumeProfileUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}